Fix HA resource restart when no replicas are running
When the helper script pacemaker_restart_bundle.sh is called
during a stack update, it restarts the pacemaker resource via
a "pcs resource restart <name>".

When all the replicas are stopped due to a previous error,
pcs won't restart them because there is nothing to stop. In
that case, one must use "pcs resource cleanup <name>".

Change-Id: I1790444d289d057e9a3f612c53efe485080978b5
Closes-Bug: #1889395