Set up right DNF module stream for Upgrades and Updates.
To ensure we get the right package versions we need to
make sure the right stream is configured for the
container-tools module.

This patch leverages the dnf module, which when passing as name
the @ identifier allows us to enable a stream for some specific
module [0], configuring the stream passed in DnfStreams heat parameter
in updates/upgrades task for step0.

If the stream passed in such parameter is alredy set, then the
module won't do anything else. On the contrary, it will disable
the current stream, enable the ones from DnfStreams and synchronize
the content.

By default the new Heat parameter defaults to empty list and it is
in the environment files where we set the right values depending
on the operation to perform (undercloud upgrade, overcloud update
or overcloud upgrade).

*****************************************************************
*BACKPORT NOTE: Add DNF module stream virt:8.2 in the overcloud *
*upgrade/update environment files.                              *
*****************************************************************

[0] - https://docs.ansible.com/ansible/latest/modules/dnf_module.html#examples
Closes-Bug: #1890631
Resolves: rhbz#1866479

Change-Id: Ib388af0a06bbb6172593dc2fd19748b770b2e335
(cherry picked from commit d3ef7c0faf1998e143dc43512ed0983d6f83df4e)