undercloud_upgrade: tear-down keepalived
Keepalived was deprecated and the VIPs are now deployed by
os-net-config.
When Keepalived is stopped, it brings down the VIPs which is
problematic since it'll remove the resources created by os-net-config
so let's teardown keepalived in the upgrade tasks here and later
during the deploy the os-net-config tool will re-create the VIPs.
Doing it at step 5 so upgrade steps which need API access still work.

Change-Id: Iee46d0394ad0b2b31da93d0a93dcfa53766fa246
Related-Bug: #1888411