Fix delegation with FreeIPA cleanup
Previously, we were delegating the IPA cleanup role to the undercloud
via localhost. This is because the keytab used to authenticate to
FreeIPA and perform the cleanup of host entries during scale down is on
the undercloud. However, when using train, ansible is invoked from the
mistral container when using `delegate_to: localhost`. In this case,
you'll end up with a privilege escalation error:

  "sudo: unable to open /run/sudo/ts/mistral: Permission denied\nsudo: a password is required\n",

This is because the mistral container doesn't have passwordless sudo,
resulting in a failed privilege escalation.

Instead, we should make sure we delegate this task to the Undercloud,
where we know the tripleo-admin user is setup properly.

Change-Id: I844f78c520d7b507d906faf7242e72dd717f9cb5
Related-Bug: 1891317