Set a higher PIDs limit for nova_libvirt container
The default limit is set to 4096. This can be reached with about 150
VMs, and therefore can lead to a situation where you're unable to start
new VMs on a compute node.

This patch integrates the modifications made by Rabi in his own
(abandonned) patch.

Note: this patch needs to be backported down to stable/train. The value
of Depends-On will need to be updated in order to point to another
patch, in paunch repository: https://review.opendev.org/747831

Change-Id: Ic414fc8826e4164ed679fbe22b82acf39c9ed7e0
Co-Authored-By: Rabi Mishra <ramishra@redhat.com>
Related: https://bugzilla.redhat.com/show_bug.cgi?id=1871885
Related-Bug: #1892817
Depends-On: https://review.opendev.org/747834