Make sure IPA has the right ACI
We need a special ACI in FreeIPA to allow etcd to obtain a
certificate with an IP SAN.  This ACI needs to be added ahead of
time.  We add a call for a validation here to make sure that the
relevant ACI has been added.

On failure, the installation will fail with instructions to add
the ACI.

Depends-On: https://review.opendev.org/#/c/749575/
Change-Id: I9baaa77b5b846c96cf075244a8ccb6889469b08e
(cherry picked from commit 32934b30abe7bc2318b4607d727f8d05f2a9e497)