ironic: stop defaulting to iscsi deploy in preparation for its deprecation
We have announced our plans to deprecate and eventually remove the iscsi
deploy interface [1]. To prepare for that, the iscsi deploy is no longer
enabled by default. The default value of IronicImageDownloadSource is changed
to 'http' to reduce the upgrade impact for clouds without swift (ironic
upstream is planning to make a similar change). Underclouds still use swift
for now to save disk space.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-September/016952.html

Change-Id: I1de6853eebc545d65ef6ce09276be3bbac2132bd