Add package install for openssl-perl
openssl-perl is used to get the directory /etc/pki/CA, which
is needed for cert requests during the deployment.  In OVB
deployments, this package is part of the image and so does not need
to be installed.

In pre-provisioned node environments, we need to document that this
package needs to be installed.  By adding this patch, we ensure that
it is there before we need it for certs.

This of course assumes we're not in some kind of airgapped environment
(which is why its dependent on IdMInstallClientPackages).  In that case,
we need to continue to doc what must be there.

Change-Id: I7fe5404144f7a75a5a1c257ceef9593719ac5dbc