[Train only] Add missing stop service steps for FFU.
We observed, when running the system_upgrade_transfer_data that few OSP
services which are still present in Train keep running before the database
backup is restored on the newly upgraded controller. This could cause some
discrepancies between the restored state and the cluster one.

Adding all the identified services present in both versions (deprecated
services will be stopped when upgrading the old nodes operating system).

Change-Id: I43e0d04866f6ce37be5de8b68aac6d43e165dc76
Closes-Bug: #1894238