Remove sync_db parameters
Currently initialization of db is implemented as an independent task
in tripleo-heat-templates and not triggered by puppet.
In puppet, all of sync db jobs are implemented by exec resources but
"exec" is not included in puppet_tags enabled, so these implementations
in puppet are never triggered.

This patch removes sync_db parameters from templates because they are
ineffective and misleading.

Change-Id: Id231c612d8ef0ebc27bf87e0b2acbb76d89c9801