Add tag's with network_data info to neutron net
Working on a tool to extract deployed networks from an existing
heat stack, adding this information so that it's available for
export.

Also when (if) we remove OS::Neutron::Net, OS::Neutron::Subnet
resources from THT storing this state in the neutron networks
will make it possible to recreate network_data in case the file
is deleted.

Depends-On: https://review.opendev.org/750662
Change-Id: I696afb8f24cc2fdeb17aaef1d3bc3a5fc22a8dcd