HA: option to use static pacemaker image name
With [1] we added the ability to use a fixed, static prefix for
the container image names that pacemaker uses to set up the HA
resources. This name is just an alias to the real tripleo
image that is being used in the control plane, and it allows to
update the namespace part of the image during a minor update
without disrupting service.

Add a new Heat parameter to use a completely static container
image name, to allow arbitrary name change during minor update,
e.g. registryA/namespaceA/imgA:tagA to registryB/namespaceB/imgB:tagB
By default, this new paramter is disabled.

[1] Id369154d147cd5cf0a6f997bf806084fc7580e01

Change-Id: I124c1e4dbcc7a8ed38079411f41a8f31c8f62284