Create post-deploy.conf on update/upgrade for octavia
Containers are restarted with new command lines during upgrade/updates
before the external_deploy_task are run that create a configuration file
that is used on the command line. This results in octavia services
failing to start.

Note: this was originally merged as:
    https://review.opendev.org/#/c/750986/
but the OctaviaBase references in templates was incorrect so the
original patch was reverted through:
    https://review.opendev.org/#/c/763561/
As the original patch had not been backported, it makes more sense to
revert and get a correct fix in place and backport that instead of
requiring backporting a series or backporting a squashed commit.

Conflicts:
  deployment/octavia/octavia-api-container-puppet.yaml

Change-Id: Ib3476e53f89b50bae72b9c95a5d3dec51ed3de7e
Related-Bug: #1863595
(cherry picked from commit 6c11a949fc0b940ebc3815aeb86156bd60df5851)