Disable snapshot support in Cinder NFS backend by default
Currently there is a known issue[1] in snapshot feature of cinder nfs
backend, which causes data corruption in several cases.
This change makes the feature disabled by default to require some
consideration by users before enabling the feature.

Note that this change makes the default value in Tripleo consistent
with the one(False) in cinder, so also fixes the inconsistency about
the default values.

[1] https://bugs.launchpad.net/cinder/+bug/1860913

Related-Bug: #1860913
Closes-Bug: #1896324
Change-Id: I12b8a01d0b28fed66be8ae0b1723dd89f6dc00ff