Force container fetch
If a user uses :latest, the container is not updated when running the
podman_image nor is it updated by podman_container. Since
podman_container will only pull if the container image doesn't exist, we
need to force the container pull here to ensure we get updated
containers as part of our deployment and updates processes.

Depends-On: https://review.opendev.org/#/c/753026/
Change-Id: Ia16cee016fb1af7b31e0f1c6da1e059221039ca4
Closes-Bug: #1895974