enable-ssh-admin: allow to override plan name
By default, the overcloud plan name is "overcloud" but it can be renamed
if the stack is named directly. Allowing the operator to override it so
they can run that script on a deployment where the stack is named
differently than the default.

Change-Id: I05029d30447592d4e16d85acc490cda69b101e9c
Closes-Bug: #1896505
(cherry picked from commit bb2351a48d4a4acf7ed933094298e27d3b91c315)