Add ctlplane network attrs to Networks resource output
Add the attributes of ctlplane network and subnets to
the output of the Networks resource.

The ctlplane attributes are added to the plan environment
by the depends-on tripleoclient change.

NOTE: This is a partial backport - it does not include setting
      'ctlplane_mtu' in group_vars.

Closes-Bug: #2017636
Depends-On: https://review.opendev.org/825689
Change-Id: Ic1772b384ad5b2003563b4e3d41a4a2f0e82f8e6
(cherry picked from commit 7e9a07212b4710077d63ae031379c1b8d9a69055)