Refactor OVNMacAddressNetwork
Move the OVNMacAddressNetwork resource out of the
Networks resource. This is a special network that
we want to create internally.

Having this network nested in the Networks resource
interfer with the work related to creating composable
networks with non-heat tooling.

Change-Id: I5551af42bc87fb8489edbfbd2abadba3dca30b0b
(cherry picked from commit 6362a0205f17d30d7a5048b38770d5ef920a3259)