Squashed backport for 'NovaAllowResizeToSameHost' parameter
We squashed the backports because the first patch wasn't right, so for
full history we backports the 2 patches in the same time.

1) Add 'NovaAllowResizeToSameHost' parameter

This parameter is used quite often but is not currently exposed in THT.
Close that gap. This is used by the API so it is that container that's
modified.

Change-Id: Idb56d7e2fd741b3d0dfaa8fd376ebe6d6ae8b867
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 0b63621a5e40d53166fe8a5625236bed82e1b3f4)

2) Fix allow_resize_to_same_host

allow_resize_to_same_host is in nova::api class, not ::nova.

Change-Id: I4cc533b3d48f3443f3470e1b25ffc39edfc9dc22
(cherry picked from commit c583e429c29ecdd88fb0c8090ce50d0b475024fa)