[Train-only] Configure podman registries during a minor update.
If a new insecure registry would be added in the update prepare
this wouldn't be included in the Podman's unsecure registries
configuration. Leading to a failed update as the overcloud nodes
wouldn't be able to reach the insecure registry.

Reusing the upgrade_taks that took care of configuring the
insecure registries, as well as adding (if it wasn't already) the
underclouds hostname in /etc/hosts. In theory, these tasks should
be idempotent. So, if all these configurations are already done the
tasks should not do anything.

Change-Id: I8a73fd31ac365146439191111bacfb4b6c5b9ba3