Fix cinder_volume upgrade tasks for stein
Prior to Train, the <service>_node_names variables are only populated
for services that have network definitions in the ServiceNetMap. In
train, this code was moved from yaql to ansible so all enabled services
get <service>_node_names defined. Additionally only docker was supported
upstream for HA, however we did support running podman when using with
RHEL8 so we need to make sure the container cli uses the expected cli.

Related-Bug: #1896537
Change-Id: If442e5b43781c1d68a126c5bcfb93ffacda78f2f