Skip yaml validation for releasenotes and zuul.d
Currently yaml validation is executed against almost all .yaml files
in tht repo, but this makes releasenote files and zuul job definitions
tested by yaml-validate.py.

This patch addes the new --skip-dir option to yaml-validate.py and
exclude these directories (and .tox which was already excluded) so that
we can ignore files which are not valid deployment templates.

Change-Id: If6fce493151ef17bfba26f8b80fc267d09408a9b