Add ssl_verify_client parameter for horizon
The recent change in puppet-horizon[1] made the ssl_verify_client
parameter mandatory when ssl_ca is set. This patch makes sure that
the ssl_verify_client parameter is set properly.

In addition, internal tls cert is not valid when internal tls is not
enabled. This patch also addresses that point, and make ssl_ca is set
only when needed.

[1] https://review.opendev.org/#/c/758041/6

Closes-Bug: #1900947
Change-Id: I286f69b8d3775d7538685e799f092ce47b5d75a7