Honor the ExternalSwiftUserTenant parameter
The ExternalSwiftUserTenant parameter is supposed to determin the
keystone project to which the swift user belong, however in fact
the parameter is not honored.

This patch ensures that the parmaeter value is used as a project name
when TripleO creates keystone resources.

Change-Id: Ib7a173643880b573c3f30b3d990c7c617231a14c
Closes-Bug: #1901652
(cherry picked from commit e618ad8d3443b3f773de52f655300ac04144b449)