node_exporter_container_image is needed in 'all' group
ceph-ansible requires the definition of the node exporter
container in all.yml, but at this time, the related variable
is scoped on the Grafana server role only.
If we want to deploy node exporter containers on all the
overcloud nodes (including the CephStorage nodes), it is
needed in the 'all' group and this is especially true if
when operators are not using the default images but satellite.

Change-Id: I7792cb252edf71ab1bb168eac1181365ca98ea7b
Closes-Bug: #1902251