Make sure octavia user belongs to octavia project
For Octavia, we have OctaviaUserName and OctaviaProjectName to define
user/project used for octavia service.
Currently tripleo creates the service project and user according to
these parameters, but the octavia user always belong to 'service'
project, not to the project defined by OctaviaProjectName.

This change ensures the octavia user belongs to the project defined by
the OctaviaProjectName parameter.

Change-Id: I32812b3cb1216c0617f3e9ccd498a2d53fec61a6