cinder/iscsi backend: remove lvm2-lvmetad.service dependency
When using iscsi/lvm backend, we create a systemd service that will help
to recreate the loop device needed by the LVM setup.

However this service was requiring lvm2-lvmetad.service to be exected on
the host but this one doesn't exist anymore in RHEL8/centos8. Let's just
remove it.

Change-Id: Id3b807955dadf8c99a9f90a96069f85c9d05e18d