[train/stein only] Add parameter to identify previous nbd-tls state
With adding UseTLSTransportForNbd feature in stein, UseTLSTransportForNbd
is enabled per default. If an environment gets upgraded where
UseTLSTransportForNbd was not enabled will make live migration to fail
as the required certs are not part of the env of the previously created
qemu process containers.

To support updating/moving from previous version not using tls-nbd
we need to be able to identify the last state from tripleoclient.
This change sets a global_config_settings parameter to identify this
be able to do this.

Related-Bug: #1900986
Related: https://bugzilla.redhat.com/show_bug.cgi?id=1888951

Change-Id: I1d5f4a714692774bac4fd42935df68a2494e46b9