Adding Ceph Dashboard to the Edge roles
This change adds the CephGrafana Service for the
Edge roles.
Also, a new DistributedComputeHCIDashboard is
introduced: the purpose of this role is to support
the StorageDashboard network on the edge site when
that network is defined and used in central site.

Change-Id: I973e1364e5f3d0e03658acdd37e316a59504825e
(cherry picked from commit 0b41751786e8c170afa73ec17853dc4a77f0b38d)