Properly compute hostname when looking for the ceph-mon container
There are some cases when the return value of the ${HOSTNAME) variable
is different from {{ ansible_hostname }}, especially when tls-everywhere
framework is involved.
This purpose of this review is to fix the FFU section that needs to know
the hostname to find the running container on the delegated host.

Change-Id: Ib269a25fcf0db66ad3144c3e20031a4da6e2e2fb
Closes-Bug: #1903537
(cherry picked from commit 4e0f61d9e2affd107b0ac9a195e23f2a71083383)