Support deploying multiple Cinder Netapp Storage backends
CinderNetappBackendName is enhanced to support a list of backend names,
and a new CinderNetappMultiConfig parameter provides a way to specify
parameter values for each backend.

The CinderNetappVolumeList parameter is properly marked as deprecated.
The corresponding parameter no longer exists in cinder and puppet-cinder,
and it was made obsolete when the CinderNetappPoolNameSearchPattern
was introduced several cycles ago.

Depends-On: I15f4828beb9db81583146fa9afdd21999ad5b078
Change-Id: Iecc11782372097bb7f24b6cbc342c38148b9c469