[train-only][ffwd] Update InstanceHA script in hybrid mode
The check-run-nova-compute script is distributed through THT which
means the new container image or rpm will not have it. This leaves
us with Queens version of the script which is python3 incompatible
and fails. We simply switch for the newer THT version of the
script before starting the Train container.

Change-Id: I389a49f724318eded04fffc83c8ecabba0e258f0
Resolves: rhbz#1897691