[train-only][ffwd] Dont reuse tripleo_step4 for hybrid state
Due to changes in paunch in Queens we had to separate out the
config of hybrid state containers. We run them up separately but
we were reusing the tripleo_step4 config id.
By creating separate config id for each hybrid container. But this
requires us to remove the running container in different config.

Change-Id: I31581f176bb4f056309f127608451caec0202a68